chore(all): [ch7233] organize components by directory
