+++ README.md
@@ -14,6 +14,7 @@ This repo provides a basic setup for developing component libraries in Vite with
 - [x] LcIcon
 - [x] LcInput
 - [x] LcModal
+- [x] LcMultiselect
 - [x] LcPagination
 - [x] LcRadioGroup & LcRadio
 - [x] LcTable
