+++ src/components/LcIcon/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LcIcon.vue'
