+++ src/components/LcInput/LcInput.d.ts
