+++ src/components/LcModal/LcModal.d.ts
